GSODRdata: climate and ecological data for the GSODR package

This data-only package was created for distributing data useful to the GSODR package and used in examples.

Source

http://chelsa-climate.org/
https://crudata.uea.ac.uk/~timm/grid/CRU_CL_2_0.html
http://maps.elie.ucl.ac.be/CCI/viewer/index.php
http://www.earthenv.org/cloud
http://www.worldclim.org

References

Karger, D. N., Conrad, O., Bohner, J., Kawohl, T., Kreft, H., Soria-Auza, R. W., et al. (2016) Climatologies at high resolution for the Earth land surface areas. arXiv preprint arXiv:1607.00217

New, M., Lister, D., Hulme, M. and Makin, I., 2002: A high-resolution data set of surface climate over global land areas. Climate Research 21:1-25 (abstract, paper)

Wilson A.M., Jetz W. (2016) Remotely Sensed High-Resolution Global Cloud Dynamics for Predicting Ecosystem and Biodiversity Distributions. PLoS Biol 14(3): e1002415. doi:10.1371/journal. pbio.1002415

Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology 25: 1965-1978.

See also