WorldClim_Bio

WorldClim_Bio

WorldClim_Bio

Format

A data frame with 26231 observations of 20 variables:

STNID

Unique station ID to link to GSOD data from GSODR

bio1

Annual Mean Temperature

bio2

Mean Diurnal Range (Mean of monthly (max temp - min temp))

bio3

Isothermality (bio2 / bio7) (* 100)

bio4

Temperature Seasonality (standard deviation * 100)

bio5

Max Temperature of Warmest Month

bio6

Min Temperature of Coldest Month

bio7

Temperature Annual Range (bio5 - bio6)

bio8

Mean Temperature of Wettest Quarter

bio9

Mean Temperature of Driest Quarter

bio10

Mean Temperature of Warmest Quarter

bio11

Mean Temperature of Coldest Quarter

bio12

Annual Precipitation

bio13

Precipitation of Wettest Month

bio14

Precipitation of Driest Month

bio15

Precipitation Seasonality (Coefficient of Variation)

bio16

Precipitation of Wettest Quarter

bio17

Precipitation of Driest Quarter

bio18

Precipitation of Warmest Quarter

bio19

Precipitation of Coldest Quarter

Source

WorlClim Global Climate Data - Free climate data for ecological modeling and GIS http://www.worldclim.org/bioclim.

WorldClim are freely available, average monthly climate data. Current conditions (interpolations of observed data, representative of 1960-1990) are freely available for download from http://www.worldclim.org/version1. Climatic elements include minimum, mean and Mean Maximum temperature and precipitation along with derived bioclimatic variables. WorldClim 1.4 (current conditions) are released under a Creative Commons Attribution-ShareAlike 4.0 International License http://creativecommons.org/licenses/by-sa/4.0/.

References

Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology 25: 1965-1978.